--- test/Circulant.jl
@@ -1,13 +1,0 @@
-function Circulant(x)
-    lx = length(x)
-    ids = [1:1:(lx-1);]
-    a = Array{Float64,2}(undef, lx,lx)
-    for i = 1:length(x)
-        if i==1
-            a[i,:] = x
-        else
-            a[i,:] = vcat(x[i],a[i-1,ids])
-        end
-    end
-    return Symmetric(a)
-end
