--- test/exponentialCorrelation.jl
@@ -1,4 +1,0 @@
-function exponentialCorrelation(x; nu = 1, length_ratio = 1)
-    R = length(x)*length_ratio
-    return exp.(-3*(x/R).^nu)
-end
